+++ zarrs/examples/custom_data_type_fixed_size.rs
@@ -22,8 +22,8 @@ use zarrs::array::{
 };
 use zarrs_data_type::{
     DataTypeExtension, DataTypeExtensionBytesCodec, DataTypeExtensionBytesCodecError,
-    DataTypeExtensionError, DataTypePlugin, FillValue, IncompatibleFillValueError,
-    IncompatibleFillValueMetadataError,
+    DataTypeExtensionError, DataTypeFillValueError, DataTypeFillValueMetadataError, DataTypePlugin,
+    FillValue,
 };
 use zarrs_metadata::{
     v3::{MetadataConfiguration, MetadataV3},
@@ -181,9 +181,8 @@ impl DataTypeExtension for CustomDataTypeFixedSize {
     fn fill_value(
         &self,
         fill_value_metadata: &FillValueMetadataV3,
-    ) -> Result<FillValue, IncompatibleFillValueMetadataError> {
-        let err =
-            || IncompatibleFillValueMetadataError::new(self.name(), fill_value_metadata.clone());
+    ) -> Result<FillValue, DataTypeFillValueMetadataError> {
+        let err = || DataTypeFillValueMetadataError::new(self.name(), fill_value_metadata.clone());
         let element_metadata: CustomDataTypeFixedSizeMetadata =
             fill_value_metadata.as_custom().ok_or_else(err)?;
         Ok(FillValue::new(element_metadata.to_ne_bytes().to_vec()))
@@ -192,12 +191,12 @@ impl DataTypeExtension for CustomDataTypeFixedSize {
     fn metadata_fill_value(
         &self,
         fill_value: &FillValue,
-    ) -> Result<FillValueMetadataV3, IncompatibleFillValueError> {
+    ) -> Result<FillValueMetadataV3, DataTypeFillValueError> {
         let element = CustomDataTypeFixedSizeMetadata::from_ne_bytes(
             fill_value
                 .as_ne_bytes()
                 .try_into()
-                .map_err(|_| IncompatibleFillValueError::new(self.name(), fill_value.clone()))?,
+                .map_err(|_| DataTypeFillValueError::new(self.name(), fill_value.clone()))?,
         );
         Ok(FillValueMetadataV3::from(element))
     }